phase-5-claud/backend/requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.1
python-multipart==0.0.22

# Database
sqlmodel==0.0.32
psycopg2-binary==2.9.11
alembic==1.14.0

# Authentication
pyjwt==2.11.0
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0

# OpenAI & Agents
openai==2.9.0
openai-agents==0.8.1
openai-agents[litellm]==0.8.1

# Data Validation
pydantic[email]==2.9.2
pydantic-settings==2.6.1

# HTTP Client
httpx==0.28.1
